--- a/.babelrc
+++ b/.babelrc
@@ -1,6 +1,6 @@
 {
   "presets": [
-    ["env", {"modules": false}]
+    ["env"]
   ],
   "plugins": [
     ["component", [{