--- a/.babelrc
+++ b/.babelrc
@@ -1,6 +1,6 @@
 {
   "presets": [
-    ["env"]
+    ["env", {"modules": false}]
   ],
   "plugins": [
     ["component", [{