--- a/deref_schema.js
+++ b/deref_schema.js
@@ -18,7 +18,7 @@ function readJsonSchema(filePath) {
 function filenameMatches(filename, allowedNames, allowedExt) {
     for (const name of allowedNames) {
         const allowedOption = name + allowedExt;
-        if (filename.toLowerCase() == allowedOption.toLowerCase()) {
+        if (filename.toLowerCase() === allowedOption.toLowerCase()) {
             return true;
         }
     }
@@ -31,40 +31,48 @@ function mergeCustomizer(objValue, srcValue) {
     }
 }
 
-if (process.argv.length != 3) {
-    throw 'The script expects a single argument: a path to directory where metadata should be stored';
-}
+function derefSchema(outputDir) {
 
-if (!config.metadataTypes) {
+  if (!config.metadataTypes) {
     throw 'Allowed metadata types not found in "clientConfig.json"';
-}
+  }
 
-const outputDir = process.argv[2];
-if (!fs.existsSync(outputDir)) {
+  if (!fs.existsSync(outputDir)) {
     fs.mkdirSync(outputDir);
-}
+  }
 
-const minMetadata = readJsonSchema(minSchemaPath);
+  const minMetadata = readJsonSchema(minSchemaPath);
 
-const metadataFiles = {}
-const dirFiles = fs.readdirSync(specializedSchemasDir);
-for (const filename of dirFiles) {
+  const metadataFiles = {};
+  const dirFiles = fs.readdirSync(specializedSchemasDir);
+  for (const filename of dirFiles) {
     const filePath = path.join(specializedSchemasDir, filename);
     const fileStat = fs.lstatSync(filePath);
     if (!fileStat.isDirectory() && filenameMatches(filename, config.metadataTypes, metadataFileExtension)) {
-        const specMetadata = readJsonSchema(filePath);
-        const mdType = specMetadata['properties']['Data_Type']['enum'][0];
-        const fullSchema = {};
-        mergeWith(fullSchema, minMetadata, specMetadata, mergeCustomizer);
-        fs.writeFileSync(path.join(outputDir, filename), JSON.stringify(fullSchema));
-        metadataFiles[mdType] = filename;
+      const specMetadata = readJsonSchema(filePath);
+      const mdType = specMetadata['properties']['Data_Type']['enum'][0];
+      const fullSchema = {};
+      mergeWith(fullSchema, minMetadata, specMetadata, mergeCustomizer);
+      fs.writeFileSync(path.join(outputDir, filename), JSON.stringify(fullSchema));
+      metadataFiles[mdType] = filename;
     }
-}
+  }
 
-if (!metadataFiles) {
+  if (!metadataFiles) {
     throw 'No metadata schemas were found';
+  }
+
+  // create unified metadata registry
+  const registryExportStmt = `export default ${JSON.stringify(metadataFiles)};`;
+  fs.writeFileSync(path.join(outputDir, metadataMappingFilename), registryExportStmt);
 }
 
-// create unified metadata registry
-const registryExportStmt = `export default ${JSON.stringify(metadataFiles)};`;
-fs.writeFileSync(path.join(outputDir, metadataMappingFilename), registryExportStmt);
+if (require.main === module) {
+  if (process.argv.length !== 3) {
+    throw 'The script expects a single argument: a path to directory where metadata should be stored';
+  }
+
+  derefSchema(process.argv[2])
+} else {
+    module.exports = derefSchema;
+}