… package.json
… src/components/SingleSelectFilter.vue
… src/components/annotation-widgets/default/MainImage.vue
… src/components/annotation-widgets/default/MainImageHeader.vue
… src/graphqlClient.ts
… src/main.ts
… src/registerElementUi.ts
… src/router.js
… src/router.ts
… src/tokenAutorefresh.ts
… tsconfig.json
… webpack.common.js
… webpack.config.js
… webpack.dev.config.js
… yarn.lock