After updating TypeScript and Webpack, a bunch of things surfaced that needed fixing or were able to be easily improved.

#### Changes:
* Lots of library updates (see `package.json`)
    * Updating webpack fixes the source map bugs and reduces code size (see Before/After below)
    * Adding `fork-ts-checker-webpack-plugin` greatly sped up compilation time
    * It turned out we weren't even using the `ajv` library - we just loaded it and loaded a schema but did all validation server-side. Removing it got rid of some webpack warnings.
    * The `assert` library didn't work in webpack'd code after the update. It was easier to remove it than fix it, as we didn't actually use most of its functionality.
* Refactor `server.js` and the `webpack.config`s so that they're easier to maintain and don't do anything unnecessary. My first attempt at removing `WebpackShellPlugin` required some async code, which meant `server.js` had to have big changes to `await` the config. It looked a lot better after the refactor, so I kept it even though I took a different path for removing `WebpackShellPlugin`.
* `deref_schema.js` is now called as a function prior to the webpack build. This means it runs synchronously instead of asynchronously. This fixes a race condition where it was possible for Webpack to try to access `metadata_schema.json` before it had finished being built. That file would also often change mid-compilation, causing redundant rebuilds during development.
* Change all routes in `router.js` so that they're named dynamic imports, so that Webpack creates more granular bundles. This pushes about 350kB of the JS code out of the initial payload into bundles that are only loaded as necessary.
    * This means a faster initial load but slightly more lag when switching to pages that weren't previously in separate bundles. There's a bug in webpack that stops it from respecting the `webpackPrefetch` hints, but someone opened a PR to solve it today so I am hopeful this will be fixed soon.

##### Before:
  Build time (dev): 38s
  Built time (prod): 58s
  Total JS (prod): 2192K
##### After:
  Build time (dev): 13s
  Built time (prod): 40s
  Total JS (prod): 2080K